+++ src/main/java/ru/ibs/recruiters_happiness/auth/db/DbApplicationUserDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.Optional;
 
 
-public interface DbApplicationUserDAO extends CrudRepository<AppUser, Integer> {
+public interface DbApplicationUserDAO extends CrudRepository<AppUser, Long> {
 
     Optional<AppUser> findByUsername(String username);
